+++ code/win32/win_local.h
@@ -111,6 +111,7 @@ extern "C" {
 #endif
 
 qboolean IN_MouseActive( void );
+qboolean IN_LegacyMouseDrivesInput( void );
 void	IN_Win32MouseEvent( int x, int y, int mstate );
 void	IN_RawMouseEvent( LPARAM lParam );